# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MusicTrackPatchPayload": {
        "additionalProperties": false,
        "description": "Body for `PUT /v1/admin/agencies/{agency_id}/music/{music_id}`.\n\nOnly ``display_name`` and ``is_default`` are editable post-upload.\n``object_key`` is owned by the upload endpoint (the binary lives on\ndisk and the key is opaque to clients) and ``duration_seconds`` is\nderived from ``ffprobe`` — both would corrupt the on-disk/DB\ninvariant if clients could rewrite them. ``extra='forbid'`` makes the\nHTTP layer return 422 when either field is supplied (feature 22).",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          }
        },
        "title": "MusicTrackPatchPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CPIHED Webhook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/agencies/{agency_id}/intro/upload": {
      "post": {
        "description": "Accepts a multipart `file` field (MP4 or MOV) and stores it under the agency's intro folder. Validates size (<=50MB) and duration (1..10s via ffprobe). Returns the asset metadata (`intro_object_key`, `intro_duration_seconds`, `intro_source = \"uploaded\"`). Errors: 422 INTRO_INVALID_MIME, 413 INTRO_FILE_TOO_LARGE, 422 INTRO_INVALID_DURATION, 404 ADMIN_AGENCY_NOT_FOUND.",
        "operationId": "upload_admin_agency_intro_v1_admin_agencies__agency_id__intro_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload the agency's intro video (multipart, MP4/MOV, <=50MB, 1-10s)",
        "tags": [
          "Admin · Intro"
        ]
      }
    },
    "/v1/admin/agencies/{agency_id}/music": {
      "get": {
        "description": "Returns every music track attached to the agency, ordered by display name. Replaces the legacy `/music-tracks` stub.",
        "operationId": "list_admin_agency_music_tracks_v1_admin_agencies__agency_id__music_get",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the agency's music library",
        "tags": [
          "Admin · Music"
        ]
      },
      "post": {
        "description": "Retired by feature 22. Use `POST /v1/admin/agencies/{id}/music/upload` (multipart) instead. Always returns 405.",
        "operationId": "register_admin_agency_music_track_v1_admin_agencies__agency_id__music_post",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "(retired) Direct metadata POST for a music track",
        "tags": [
          "Admin · Music"
        ]
      }
    },
    "/v1/admin/agencies/{agency_id}/music/upload": {
      "post": {
        "description": "Accepts a multipart body with `file` (audio binary), `display_name` (label, 1..200 chars) and `is_default` (`true`/`false`). Supported MIME types: audio/mpeg, audio/mp4, audio/wav. The backend persists the binary under the agency's music folder, probes the duration via ffprobe and registers a row in `agency_music_tracks`. Returns 201 with the saved `music_track`. Errors: 400 MUSIC_TRACK_AUDIO_INVALID (bad MIME / ffprobe failure / duration > 10min), 413 if the payload is over 20 MB, 422 if `display_name` is missing or blank, 404 if the agency is unknown.",
        "operationId": "upload_admin_agency_music_track_v1_admin_agencies__agency_id__music_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a new music track (multipart, <=20MB)",
        "tags": [
          "Admin · Music"
        ]
      }
    },
    "/v1/admin/agencies/{agency_id}/music/{music_id}": {
      "delete": {
        "operationId": "decommission_admin_agency_music_track_v1_admin_agencies__agency_id__music__music_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "music_id",
            "required": true,
            "schema": {
              "title": "Music Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decommission a music track from the agency library",
        "tags": [
          "Admin · Music"
        ]
      },
      "get": {
        "operationId": "inspect_admin_agency_music_track_v1_admin_agencies__agency_id__music__music_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "music_id",
            "required": true,
            "schema": {
              "title": "Music Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect a single music track",
        "tags": [
          "Admin · Music"
        ]
      },
      "put": {
        "description": "Updates the descriptive fields of a music track. Omitted fields preserve the previously stored value.",
        "operationId": "reconfigure_admin_agency_music_track_v1_admin_agencies__agency_id__music__music_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "music_id",
            "required": true,
            "schema": {
              "title": "Music Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MusicTrackPatchPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconfigure an existing music track",
        "tags": [
          "Admin · Music"
        ]
      }
    },
...
    "/v1/admin/agencies/{agency_id}/outro/upload": {
      "post": {
        "description": "Accepts a multipart `file` field (MP4 or MOV) and stores it under the agency's outro folder. Validates size (<=50MB) and duration (1..10s via ffprobe). Returns the asset metadata (`outro_object_key`, `outro_duration_seconds`, `outro_source = \"uploaded\"`). Errors: 422 OUTRO_INVALID_MIME, 413 OUTRO_FILE_TOO_LARGE, 422 OUTRO_INVALID_DURATION, 404 ADMIN_AGENCY_NOT_FOUND.",
        "operationId": "upload_admin_agency_outro_v1_admin_agencies__agency_id__outro_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload the agency's outro video (multipart, MP4/MOV, <=50MB, 1-10s)",
        "tags": [
          "Admin · Outro"
        ]
      }
    },
...
  }
}